Correções de problemas encontrados

+++ toCSV.php
@@ -32,9 +32,11 @@ try {
 }
 
 try {
-    $logger->info('Apagando conteúdo original...');
-    $oCSV = new Directory($outputCSV);
-    $oCSV->recursive()->delete();
+    if (file_exists($outputCSV)) {
+        $logger->info('Apagando conteúdo original...');
+        $oCSV = new Directory($outputCSV);
+        $oCSV->recursive()->delete();
+    }
 } catch (Exception $ex) {
     $logger->notice($ex->getMessage());
 }
@@ -54,32 +56,31 @@ try {
 
 try {
     $logger->info('Gerando arquivo LIQUIDACAO...');
-    
+
     $empenhoPath = new Path($outputCSV, 'EMPENHO.csv');
     $empenhoHandle = fopen($empenhoPath->getRealPath(), 'r');
-    if($empenhoHandle === false){
+    if ($empenhoHandle === false) {
         throw new ErrorException("Falha ao abrir {$empenhoPath->getPath()}");
     }
     $empenho = new DataFrame(new CSVReader($empenhoHandle, ';', true));
-    
+
     $liquidacPath = new Path($outputCSV, 'LIQUIDAC.csv');
     $liquidacHandle = fopen($liquidacPath->getRealPath(), 'r');
-    if($liquidacHandle === false){
+    if ($liquidacHandle === false) {
         throw new ErrorException("Falha ao abrir {$liquidacPath->getPath()}");
     }
     $liquidac = new DataFrame(new CSVReader($liquidacHandle, ';', true));
-    
+
     $liquidacaoAssembler = new LiquidacaoAssembler($liquidac, $empenho);
     $dfLiquidacao = $liquidacaoAssembler->assemble();
-    
+
     $liquidacaoOutput = new Path($outputCSV, 'LIQUIDACAO.csv');
     $liquidacaoHandle = fopen($liquidacaoOutput->getPath(), 'w');
-    if($liquidacaoHandle === false){
+    if ($liquidacaoHandle === false) {
         throw new ErrorException("Falha ao abrir {$liquidacaoOutput->getPath()}");
     }
     $liquidacaoWriter = new CSVWriter2($dfLiquidacao, $liquidacaoHandle, ';', true);
     $liquidacaoWriter->write();
-
 } catch (Exception $ex) {
     $logger->emergency($ex->getMessage());
     exit($ex->getCode());
@@ -87,32 +88,31 @@ try {
 
 try {
     $logger->info('Gerando arquivo PAGAMENTO...');
-    
+
     $empenhoPath = new Path($outputCSV, 'EMPENHO.csv');
     $empenhoHandle = fopen($empenhoPath->getRealPath(), 'r');
-    if($empenhoHandle === false){
+    if ($empenhoHandle === false) {
         throw new ErrorException("Falha ao abrir {$empenhoPath->getPath()}");
     }
     $empenho = new DataFrame(new CSVReader($empenhoHandle, ';', true));
-    
+
     $pagamentPath = new Path($outputCSV, 'PAGAMENT.csv');
     $pagamentHandle = fopen($pagamentPath->getRealPath(), 'r');
-    if($pagamentHandle === false){
+    if ($pagamentHandle === false) {
         throw new ErrorException("Falha ao abrir {$pagamentPath->getPath()}");
     }
     $pagament = new DataFrame(new CSVReader($pagamentHandle, ';', true));
-    
+
     $pagamentoAssembler = new PagamentoAssembler($pagament, $empenho);
     $dfPagamento = $pagamentoAssembler->assemble();
-    
+
     $pagamentoOutput = new Path($outputCSV, 'PAGAMENTO.csv');
     $pagamentoHandle = fopen($pagamentoOutput->getPath(), 'w');
-    if($pagamentoHandle === false){
+    if ($pagamentoHandle === false) {
         throw new ErrorException("Falha ao abrir {$pagamentoOutput->getPath()}");
     }
     $pagamentoWriter = new CSVWriter2($dfPagamento, $pagamentoHandle, ';', true);
     $pagamentoWriter->write();
-
 } catch (Exception $ex) {
     $logger->emergency($ex->getMessage());
     exit($ex->getCode());
@@ -123,7 +123,7 @@ try {
     $latestPath = new Path(dirname($outputCSV), 'latest');
     $latestDir = new Directory($latestPath->getPath());
     $latestDir->delete();
-    
+
     $logger->info('Copiando conteúdo para latest...');
     $output = new Directory($outputCSV);
     $output->copy($latestPath->getPath());
